% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bsef.R
\name{get_bsef_data}
\alias{get_bsef_data}
\title{Get Bloomberg SEF data}
\usage{
get_bsef_data(date, asset_class = NULL)
}
\arguments{
\item{date}{the date for which data is required as Date or DateTime
object. Only the year, month and day elements of the object are used. Must
be of length one.}

\item{asset_class}{the asset class for which you would like to download
trade data. Valid inputs are \code{"CR"} (credit), \code{"IR"} (rates),
\code{"EQ"} (equities), \code{"FX"} (foreign exchange), \code{"CO"}
(commodities). Can be a vector of these.}
}
\value{
a data frame containing the requested data, or an empty data frame
if data is unavailable
}
\description{
The Bloomberg Swap Execution Facility (SEF) offers customers the ability to
execute derivative instruments across a number of different asset classes. It
is required to make publicly available price, trading volume and other trading
data. It publishes this data on its website. I have reverse engineered the
JavaScript libraries used by its website to call the Bloomberg Application
Service using \code{POST} requests to a target URL. The \code{POST} calls
are done by asset class. This function iterates across all asset classes
(credit, equities, foreign exchange, rates and commodities).
}
\examples{
library (lubridate)
# All asset classes
get_bsef_data(ymd(20140528))
# Only IR and FX asset classes
get_bsef_data(ymd(20140528), c("IR", "FX"))
}
\references{
\href{http://data.bloombergsef.com}{Bloomberg SEF data}
}

