% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ddr.R
\name{get_ddr_data}
\alias{get_ddr_data}
\title{Get DDR data}
\usage{
get_ddr_data(date, asset_class, clean = TRUE)
}
\arguments{
\item{date}{the date for which data is required as Date or DateTime
object. Only the year, month and day elements of the object are used.}

\item{asset_class}{the asset class for which you would like to download
trade data. Valid inputs are \code{"CR"} (credit), \code{"IR"} (rates),
\code{"EQ"} (equities), \code{"FX"} (foreign exchange), \code{"CO"}
(commodities).}

\item{clean}{where or not to clean up temporary files that are created
during this process. Defaults to \code{TRUE}.}
}
\value{
a \code{tbl_df} that contains the requested data.
}
\description{
The DTCC Data Repository is a registered U.S. swap data repository that allows
market participants to fulfil their public disclosure obligations under
U.S. legislation. This function will give you the ability to download
trade-level data that is reported by market participants. The field names
are (and is assumed to be) the same for each asset class.
}
\examples{
\dontrun{
library("lubridate")
get_ddr_data(ymd(20140430), "IR")
}
}
\references{
\href{https://rtdata.dtcc.com/gtr/}{DDR Real Time Dissemination Platform}
}

