% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/icap.R
\name{get_icap_data}
\alias{get_icap_data}
\title{Get ICAP (IGDL & ICAP US) SEF data}
\usage{
get_icap_data(date, clean = TRUE)
}
\arguments{
\item{date}{the date for which data is required as Date or DateTime
object. Only the year, month and day elements of the object are used. Must
be of length one.}

\item{clean}{flag indicating whether ICAP temp files should be cleaned
(default: \code{TRUE})}
}
\value{
a data frame containing the requested data, or an empty data frame
if data is unavailable
}
\description{
ICAP offers customers the ability to execute derivative instruments on two
SEF faciltiies: one domiciled in the UK (IGDL) and the other domiciled in the
US (ICAP US). IGDL offers G3 rates products (USD, EUR and GBP - IRS, OIS,
FRAs, IROs, Swaptions, Inflation and Exotic Swaps and Options) only as the
Made Available for Trade (MAT) determinations create the greatest need for a
cross-border solution for many instruments within this product set. All other
products and instruments will continue to be serviced on ICAP's existing US
SEF, ICAP SEF (US) LLC. Irrespectively of the entity, is required to make
publicly available price, trading volume and other trading
data. It publishes complete historical data sets in two zip files: one for
IGDL and the other for ICAP US. These are downloaded, unzipped and data for
the date requested is formatted and returned to you.
}
\examples{
\dontrun{
library (lubridate)
get_icap_data(ymd(20140528))
}
}
\references{
\href{http://www.icap.com/what-we-do/global-broking/sef.aspx}{ICAP SEF}
}
\keyword{internal}

