% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_bsef_data}
\alias{get_bsef_data}
\title{Get Bloomberg SEF data}
\usage{
get_bsef_data(date)
}
\arguments{
\item{date}{the date for which data is required as Date or DateTime
object. Only the year, month and day elements of the object are used.}
}
\value{
a data frame containing the requested data, or an empty data frame
if data is unavailable
}
\description{
The Bloomberg Swap Execution Facility (SEF) offers customers the ability to
execute derivative instruments across a number of different asset classes. It
is required to make publicly available price, trading volume and other trading
data. It publishes this data on its website. I have reverse engineered the
JavaScript libraries used by its website to call the Bloomberg Application
Service using \code{POST} requests to a target URL. The \code{POST} calls
are done by asset class. This function iterates across all asset classes
(credit, equities, foreign exchange, rates and commodities).
}
\references{
\href{http://data.bloombergsef.com}{Bloomberg SEF data}
}

