% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_bsef_data}
\alias{get_bsef_data}
\title{Get Bloomberg SEF data}
\usage{
get_bsef_data(date, asset_class)
}
\arguments{
\item{date}{the date for which data is required as Date or DateTime
object. Only the year, month and day elements of the object are used.}

\item{asset_class}{the derivatives asset class for which data is required.
Must be one of the following (case-invariant): \code{'CR'} (credit),
\code{'EQ'} (equity), \code{'FX'} (foreign exchange), \code{'IR'} (interest
rates), \code{'CO'} (commodities).}
}
\value{
a data frame containing the requested data, or if no data is available,
\code{NULL}.
}
\description{
The Bloomberg Swap Execution Facility (SEF) offers customers the ability to
execute derivative instruments across a number of different asset classes. It
is required to make publicly available price, trading volume and other trading
data. It publishes this data on its website. I have reverse engineered the
JavaScript libraries used by its website to call the Bloomberg Application
Service using \code{POST} requests to a target URL.
}

